Verify User method

+++ src/backend/src/main/java/com/sprintership22/backend/service/UserServiceImplementation.java
@@ -13,18 +13,29 @@ public class UserServiceImplementation implements UserService {
 	
 	@Override
 	public User saveUser(User user) {
+		if (verifyUser(user))
+		{
+			return null;
+		}
+		
 		return userRepository.save(user);
 	}
 
 	@Override
-	public void deleteUser(User user) {
+	public boolean deleteUser(User user) 
+	{
+		if (!verifyUser(user))
+		{
+			return false;
+		}
+		
 		userRepository.delete(user);
+		return true;
 	}
 
 	@Override
 	public boolean verifyUser(User user) {
-		//userRepository.getSingleResult();
-		return false;
+		return userRepository.existsById(user.getEmployeeID());
 	}
 	
 }
